% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexCalculation.R
\name{brier_score}
\alias{brier_score}
\title{Brier Score}
\usage{
brier_score(actual, predicted_probs)
}
\arguments{
\item{actual}{Vector of observed outcomes (0/1).}

\item{predicted_probs}{Vector of predicted probabilities.}
}
\value{
Numeric Brier score.
}
\description{
Measures mean squared error of probabilistic predictions.
}
