% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexCalculation.R
\name{Fbp}
\alias{Fbp}
\title{Fbp: Modified F1 Score}
\usage{
Fbp(actual, predicted)
}
\arguments{
\item{actual}{Vector of observed binary outcomes.}

\item{predicted}{Vector of predicted binary outcomes.}
}
\value{
Numeric Fbp score.
}
\description{
Computes Fbp (Fowlkes-F1) score from binary predictions.
}
