% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexCalculation.R
\name{fowlkes_mallows}
\alias{fowlkes_mallows}
\title{Fowlkes-Mallows Index}
\usage{
fowlkes_mallows(actual, predicted_binary)
}
\arguments{
\item{actual}{Vector of observed binary outcomes.}

\item{predicted_binary}{Vector of predicted binary outcomes.}
}
\value{
Numeric FM index.
}
\description{
Computes Fowlkes-Mallows index for binary predictions.
}
