% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAAPoints.R
\name{generateAAPoints}
\alias{generateAAPoints}
\title{Generate Artificial Absence Points for SDM Evaluation}
\usage{
generateAAPoints(aa_mask, nPoints)
}
\arguments{
\item{nPoints}{Integer. Number of artificial absence points to generate.}

\item{rasters}{A \code{terra::SpatRaster} object containing environmental covariates.}

\item{presence}{An \code{sf} object of known presence locations used for testing.}
}
\value{
An \code{sf} object with sampled artificial absence points.
}
\description{
Generates artificial absence (AA) points for presence-artificial-absence (PAA) evaluation of species distribution models.
Uses Area of Applicability (AOA) to identifiy regions that have a high dissimilarity compared ot presence points.
}
