% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexCalculation.R
\name{mcc}
\alias{mcc}
\title{Matthews Correlation Coefficient (MCC)}
\usage{
mcc(actual, predicted_binary)
}
\arguments{
\item{actual}{Vector of observed binary outcomes.}

\item{predicted_binary}{Vector of predicted binary outcomes.}
}
\value{
Numeric MCC value.
}
\description{
Computes MCC for binary classification.
}
