% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexCalculation.R
\name{omission}
\alias{omission}
\title{Omission Rate}
\usage{
omission(actual, predicted)
}
\arguments{
\item{actual}{Vector of observed binary outcomes.}

\item{predicted}{Vector of predicted binary outcomes.}
}
\value{
Numeric omission rate.
}
\description{
Computes omission rate for binary predictions.
}
