% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexCalculation.R
\name{orss}
\alias{orss}
\title{Odds Ratio Skill Score (ORSS)}
\usage{
orss(actual, predicted)
}
\arguments{
\item{actual}{Vector of observed binary outcomes.}

\item{predicted}{Vector of predicted binary outcomes.}
}
\value{
Numeric ORSS value.
}
\description{
Evaluates prediction skill based on contingency table odds ratio.
}
