% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceEstimation.R
\name{performanceEstimation}
\alias{performanceEstimation}
\title{SDM performance evaluation}
\usage{
performanceEstimation(
  prediction,
  presence = NA,
  absence = FALSE,
  background = TRUE,
  aa = TRUE,
  environmentalVariables = NA,
  noPointsTesting = NA,
  replicates = 100
)
}
\arguments{
\item{prediction}{A \code{terra::SpatRaster} object with the prediction map.}

\item{presence}{Required. A \code{sf} object representing known presence locations.}

\item{absence}{Optional. A \code{sf} object representing known absence locations. If not provided, \code{background} or \code{aa} must be used.}

\item{background}{Optional. Either a \code{sf} object with background points, or logical \code{TRUE} to auto-generate them. Logical \code{FALSE} to not calculate them}

\item{aa}{Optional. A \code{sf} object representing artificial absence points. If not provided, they are derived using AOA (Area of Applicability).}

\item{environmentalVariables}{Optional unless \code{background = TRUE} or \code{aa} is missing. A \code{terra::SpatRaster} with environmental covariates.}

\item{noPointsTesting}{Integer. Number of background or artificial absence points to generate.}
}
\value{
A named \code{list} with the following components:
\describe{
\item{indexPA}{Performance metrics using provided absence data.}
\item{indexPAA}{Performance metrics using artificially generated absence data (AOA-based).}
\item{indexPBG}{Performance metrics using background data.}
}
}
\description{
Evaluates the performance of species distribution models (SDMs).
The evaluation supports three data types: presence-absence (PA), presence-artificial-absence (PAA), and presence-background (PBG).
Metrics include predictive accuracy (AUC, COR), predictive error measures (MAE, BIAS), across cross-validation replicates.
}
\details{
The index combines accuracy metrics and a spatial correlation stability metric across folds to
estimate the reliability and generalizability of the model. COR and AUC measure predictive performance,
while layer-wise correlation (Kappa) measures spatial consistency across folds.
}
\examples{
\dontrun{
  result <- evaluateSDMPerformance(
              prediction = prediction_raster,
              presence = presence_points,
              background = TRUE,
              environmentalVariables = env_rasters,
              replicates = 50)
  result$indexPA
  result$indexPAA
  result$indexPBG
}

}
