% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexCalculation.R
\name{sedi}
\alias{sedi}
\title{Symmetric Extremal Dependence Index (SEDI)}
\usage{
sedi(actual, predicted, eps = 1e-10)
}
\arguments{
\item{actual}{Vector of observed binary outcomes.}

\item{predicted}{Vector of predicted binary outcomes.}

\item{eps}{Small number to prevent division by zero. Default = 1e-10.}
}
\value{
Numeric SEDI value.
}
\description{
Evaluates prediction skill for rare events.
}
