% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexCalculation.R
\name{tcr}
\alias{tcr}
\title{Top-q\% Capture Rate}
\usage{
tcr(predicted, observed, prediction, q = 0.1)
}
\arguments{
\item{predicted}{Vector of predicted values.}

\item{observed}{Vector of observed binary outcomes.}

\item{prediction}{\code{terra::SpatRaster} object used to compute threshold.}

\item{q}{Numeric. Top proportion threshold (default 0.10 = top 10\%).}
}
\value{
Numeric proportion of presences captured.
}
\description{
Computes proportion of presences captured in top q\% of predicted values.
}
