% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMetrics.R
\name{calculateMetrics}
\alias{calculateMetrics}
\title{Calculate Performance Metrics for a single replicate/model/map}
\usage{
calculateMetrics(prediction, presence, absence_or_bg_sf)
}
\arguments{
\item{prediction}{A \code{terra::SpatRaster} object containing model predictions.}

\item{presence}{An \code{sf} object of known presence locations.}

\item{absence_or_bg_sf}{An \code{sf} object of absence, background, or artificial absence points.}
}
\value{
A \code{data.frame} with calculated metrics for the replicate.
}
\description{
Calculates a suite of performance metrics for a single replicate of SDM evaluation.
Supports presence-background (PBG), presence-artificial-absence (PAA), and presence-absence (PA) datasets.
}
